backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
openai>=1.0.0
pillow>=10.0.0
aiofiles>=23.0.0
python-telegram-bot>=20.0.0
qrcode>=7.0
httpx>=0.25.0
rapidfuzz>=3.0.0
tenacity>=8.0.0
openpyxl>=3.1.2
reportlab>=4.0.0
duckduckgo_search>=4.0.0
